Fix a bunch of stuff. Almost works correctly.

+++ api/src/main/java/org/spout/api/math/IntVector3.java
@@ -102,7 +102,7 @@ public class IntVector3 extends IntVector2 {
 	}
 
 	public Vector3 toVector3() {
-		return new Vector3((float) getX(), (float) getY(), (float) getZ());
+		return new Vector3(getX(), getY(), getZ());
 	}
 
 	@Override
